PMM-7574 move WarningBlock to shared
