+++ public/app/percona/backup/components/StorageLocations/RemoveStorageLocationModal/RemoveStorageLocationModal.tsx
@@ -1,6 +1,6 @@
 import { DeleteModal } from 'app/percona/shared/components/Elements/DeleteModal';
 import React, { FC } from 'react';
-import { WarningBlock } from '../WarningBlock/WarningBlock';
+import { WarningBlock } from '../../../../shared/components/Elements/WarningBlock/WarningBlock';
 import { Messages } from './RemoveStorageLocationModal.messages';
 import { RemoveStorageLocationModalProps } from './RemoveStorageLocationModal.types';
